Convenience Tools, LanguageServerFixes and wpa_ctrl
Added convenience tools to synchronize the target libraries with the
libaries in platform for cross compilation. Fixed problem with
LanguageServer, where it couldn't find certain headers which were
outside of the sysroot. Removing this flag from the
compile_commands.json file resolved the issue -> Added commands to do
that in CMakeLists.txt after the program is built. Added wpa_ctrl which
will be used to control wpa_supplicant.